requirements.txt
pandas
pycryptodome
matplotlib